cleaned up __init__.py: seperation of concerns
